StringGen/requirements-dev.txt
# Development dependencies for SCA String Generator
# Install with: pip install -r requirements-dev.txt

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-benchmark>=4.0.0

# Code quality
black>=23.0.0
pylint>=2.17.0
mypy>=1.0.0
flake8>=6.0.0

# Documentation
pyyaml>=6.0  # For YAML config file support

# Performance profiling
memory-profiler>=0.60.0
line-profiler>=4.0.0

# Optional: For distributed generation (future enhancement)
# mpi4py>=3.1.0